# Add Rack::Deflater for compression